Fix: evaluate the block argument when used.
Actually I dislike this approach as it seems we are using the mixin
dynamically scoped, instead of capturing one defined when the block
argument is itself defined.

But this is the behavior of Pug e.g. in https://github.com/hypered/design/blob/main/struct/templates/prototypes/refli/blog-index.pug.

The example code is similar to that page:

blog-index extends page, which includes a mixin +arrow-right.

extends-mixins extends body, which includes a mixin +list.

From another point of view, if we consider the extends as a form of
include, maybe it makes sense for it to bring into scope the definition
of the mixin.