Move the Inline module code in Parse and Evaluate.
I was hoping the Inline module, implementing basic string interpolation
using similar code to the `template` library, could be kept separate.

But the string appearing in a placeholder, e.g. #{name}, can actually be
a more general expression, e.g. #{dict['name']}. This means we want the
Var Text case be actually Var Code, and be able to reuse our Code parser
when we parse placeholders.

So we remove the Inline module, and instead move the data type
definition for Inline to Syntax, the parsing code to Parse, and the
evaluation code (the `render` function) to Evaluate.