Fix: don't evaluate a fragment definition body.
The new example where we call `content` instead of `block content` was
failing before this patch as `content` was not yet defined. In the block
case, when the argument is not found, there is a default content so the
problem was not visible.

We also add a call to preProcessNodes in the PugFragmentCall case. This
is not related to the fix.