Keep the original AST.
At some point, we decided to return a simplified (flattened) AST (e.g.
only return the taken branch of PugIf). But we want to be able to
introspect the AST for instance to visualize where some fragment
definitions are used. So we start reverting this simplification and make
it optional.