Fragments passed as args can refer to each other.
This is shown in the new example with the content passed to `post` that
now can refer to `g`.

This also shows (because we had to rename "content" to something else,
in this case "second") that we have problems with the fact that we don't
declare formal fragment parameters. The content was supposed to be given
as an argument to `imports`, but here we have defined `div` as a
fragment that also takes a `content` argument.